refactor: StrikeBallEvalutor로 모듈 분리
