+++ seokhwan/BaseballGame/BaseballGame/BaseballGame.swift
@@ -1,13 +1,7 @@
 import Foundation
 
 struct BaseballGame {
-    enum StrikeBallResult {
-        case correct
-        case partialCorrect(strikes: Int, balls: Int)
-        case incorrect
-    }
-
-    var attempts = [Int]()
+    private var attempts = [Int]()
 
     mutating func startApp() {
         while true {
@@ -49,7 +43,7 @@ struct BaseballGame {
                 continue
             }
 
-            switch evaluate(answer: answer, from: correctAnswer) {
+            switch StrikeBallEvaluator.evaluate(answer: answer, from: correctAnswer) {
             case .correct:
                 print("정답입니다!\n")
                 attempts.append(attempt)
@@ -77,22 +71,6 @@ struct BaseballGame {
               answer.isDistinct else { return nil }
 
         return answer
-    }
-
-    func evaluate(answer: Int, from correctAnswer: Int) -> StrikeBallResult {
-        guard answer != correctAnswer else { return .correct }
-        let answer = answer.digits
-        let correctAnswer = correctAnswer.digits
-
-        var strikes = 0
-        var balls = 0
-
-        answer.enumerated().forEach { index, _ in
-            strikes += answer[index] == correctAnswer[index] ? 1 : 0
-            balls += answer[index] != correctAnswer[index] && correctAnswer.contains(answer[index]) ? 1 : 0
-        }
-
-        return strikes == 0 && balls == 0 ? .incorrect : .partialCorrect(strikes: strikes, balls: balls)
     }
 
     func printGameRecords() {
